resource: add metadata for elastic job type To enable grow and shrink operations for elastic jobs, additional metadata attached to each resource graph node is necessary.  By default all jobs will have the jobtype attribute set to "static," but this can be set to "elastic" for elastic resource requests.
If a vertex is allocated to multiple jobs,
the jobid must be associated with the
corresponding jobtype in the schedule_t
struct.  Note that reservations is extended
to use the spantype_t struct in case that
is useful in the future.  Keeping a counter
of the elastic jobs allocated to a vertex will
also be necessary for grow/shrink.