README: remove module structure tangent in installation instructions
Given recent changes to the README, this paragraph now comes across as a
tangent rather than a useful description of how our modules and codebase
are structured.  The first section of the README does a good enough job
of describing those things already, that I think it is worth deleting
this paragraph.