writer: fix a type conversion error
Problem: The code converts 32-bit unsigned int
into a 64-bit integer for Jasson packing.
As a result, it reads 4 more random bytes around
the "needs" variable.
Depending on the memory layout, this nondeterministically
writes out a completely incorrect integer value
for the size field of a resource vertex.

Add an explict type casting so that the compiler
promotes 32-bit integer into 64-bit integer
before it is being read by Jansson.