testsuite: add tests to check availability based on status
add tests to resource-status tests to ensure that setting
resource state results in expected allocation behavior.
Namely, resources marked down should be unavailable,
and those marked down and then up should be available.
Additionally, add tests for allocation, allocation
or reservation, and allocation with satisfiability
for a cluster whose resources are marked down and up.
As a sanity check, add a test to reject an unsatisfiable
request regardless of whether a node is marked down.