resource-query: promote attach to production
Problem: without ctors, dtors, and
assignment overloads for planner and
planner_multi, Boost vecS resizing called
the schedule_t and pool_infra_t which
could not deep copy the structs. As
a result, the resizing caused all
allocations and reservations to be
destroyed. A check in resource-query
ensured resource graph mutation
could not be performed via `attach`
on a resource graph with allocations
or reservations.

Remove this restriction.