resource: add support for post partial cancel processing
Problem: a partial cancel successfully removes the allocations of the
other resource vertices (especially core, which is installed in all
pruning filters by default) because they have broker ranks. However,
when the final .free RPC fails to remove an ssd vertex allocation the
full cleanup cancel exits with an error when it hits the vertices it
has already cancelled.

Add support for treating the condition as normal behavior if it is
encountered post partial cancel.