reapi: expose shrink (remove_subgraph)
Problem: a full elasticity implementation for the go
bindings (and other bindings) requires an ability to shrink.
The function is implemented (remove_subgraph in the reapi
implementation c++) but not exposed.
Solution: expose the function via a corresponding "shrink"
API call. This possibly does not account for proper
cancel of jobs on those resources, but should suffice
for testing of the endpoint.

Signed-off-by: vsoch <vsoch@users.noreply.github.com>