evaluators: add extraction of jobid and agfilter bool
Problem: a jobid and aggregate filter boolean are now supported in the
evaluator and validator. The evaluator needs support for returning the
parsed jobid and agfilter bool to the traverser, so it can change
behavior based on whether a jobid or agfilter boolean is present in the
evaluation string.

Add the capability to extract the jobid as a uint64_t and agfilter as a
bool and return them to the client.