qmanager: avoid requirement of flux-core > 0.69.0
Problem: the raw RPC interfaces are changing to use a size_t
instead of int for payload size.

Earlier qmanager was changed to use size_t.  It didn't occur to
me at the time, but we can avoid using the raw RPC interface entirely
since all payloads are strings.  This allows fluxion to continue to
work with older versions.

Do that instead.