qmanager: fix partial hello for match-format=rv1
Problem: the scheduler re-allocates the entire resource set
for a partially released job when match-format=rv1

libschedutil provides a diminished Rv1 to the "hello" callback that is
suitable for match-format=rv1_nosched, but since the scheduling key
(JGF) is passed through unchanged, the full resource set is reallocated
when match-format=rv1.

Add a queue->remove() on the free ranks to compensate.

Co-authored-by: Daniel Milroy <daniel.milroy@gmail.com>