readers: add JGF support for updating only non-freed ranks
Problem: the JGF reader doesn't support excluding previously freed ranks
during update allocate. This leads to state divergence between sched
and core because Fluxion uses the initial, full JGF to reconstruct an
allocation state.

Add support for identifying and skipping vertices and edges with ranks
corresponding to those freed in previous partial cancellations. The JGF
reader unpacks the `free_ranks` key inserted in qmanager in the
`scheduling` key payload and uses the idset to exclude vertices and
edges from the JGF when fetching and updating vertices and edges.