traversers: ensure implicit exclusivity behaves like explicit
Problem: there is unexpected and unintentional behavior of the implicit
exclusivity of resources specified below a slot in a jobspec. In
particular, while the vertices are allocated or reserved and their
metadata entered appropriately, their counts are not propagated to
parent aggregate filters. The missing propagation is due to exclusive
being set to false in `dfu_impl_t::prune()`.

This behavior translates to aggregate filters tracking fewer than the
actual allocated or reserved counts. This condition negates the benefit
of aggregate filters and leads to searching far more of the graph that
necessary.

Add a jobspec traversal in prime_jobspec. The traversal sets all
resources under each slot as exclusive. It also enables extension to
allowing slots to contain resources requested non-exclusively.