test: exclusivity check
Problem: we need to be able to check that a change to fluxion does not
change cpu mapping decisions. This is a WIP set of changes that will
discovery the local resources with hwloc, and then submit simple
jobs and compare the discovered mapping with expected, where expected
is a consistent pattern we can derive. The identifiers of the cpusets
should hypothetically not matter as long as the patterns are consistent.
Finally, I tested using faux or emulated resources, and everything works
up until when we need to get the cpusets via the pid. I was not able to
get a pid that makes sense. That said, it will be worth one more try
before giving up!

Signed-off-by: vsoch <vsoch@users.noreply.github.com>